we got semi soc working

+++ firmware/quintuna/VC/src/io/io_lowVoltageBattery.h
@@ -34,7 +34,7 @@ bool io_lowVoltageBattery_init();
  *
  * @return SOC percentage on success, or -1.0f on error.
  */
-float io_lowVoltageBattery_get_SOC();
+double io_lowVoltageBattery_get_SOC();
 
 /**
  * @brief Gets the battery voltage.
@@ -44,7 +44,7 @@ float io_lowVoltageBattery_get_SOC();
  * @return The battery voltage on success, or -1 on error.
  */
  uint16_t io_lowVoltageBattery_get_voltage(voltage_cmd_t voltage_cell);
- 
+
 // intial setup for battery monitoring chip
 bool io_lowVoltageBattery_initial_setup(void);
